+++ src/main/java/cart/db/dao/CouponDao.java
@@ -1,6 +1,6 @@
-package cart.persistance.dao;
+package cart.db.dao;
 
-import cart.persistance.entity.CouponEntity;
+import cart.db.entity.CouponEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
