+++ src/main/java/cart/db/entity/OrderItemEntity.java
@@ -1,4 +1,4 @@
-package cart.persistance.entity;
+package cart.db.entity;
 
 import cart.domain.product.OrderItem;
 import cart.domain.product.Product;
