+++ src/main/java/cart/repository/CouponRepository.java
@@ -4,8 +4,8 @@ import cart.domain.coupon.Coupon;
 import cart.domain.discountpolicy.DiscountPolicy;
 import cart.domain.discountpolicy.DiscountPolicyProvider;
 import cart.domain.discountpolicy.DiscountType;
-import cart.persistance.dao.CouponDao;
-import cart.persistance.entity.CouponEntity;
+import cart.db.dao.CouponDao;
+import cart.db.entity.CouponEntity;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
