+++ src/main/java/cart/repository/ProductRepository.java
@@ -1,16 +1,11 @@
 package cart.repository;
 
 import cart.domain.product.Product;
-import cart.persistance.dao.ProductDao;
-import cart.persistance.entity.ProductEntity;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
+import cart.db.dao.ProductDao;
+import cart.db.entity.ProductEntity;
 import org.springframework.stereotype.Repository;
 
-import java.sql.PreparedStatement;
-import java.sql.Statement;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Repository
